# services/booking-service/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API para reservas, asientos, checkout y tickets.",
        "title": "SeatGuard Booking Service API",
        "contact": {
            "name": "Lucas Cabral",
            "email": "lucassimple@hotmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:4000",
    "basePath": "/api/v1",
    "paths": {
        "/booking-order": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Obtiene todos los registros de booking orders",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "booking-order"
                ],
                "summary": "Obtener todos los booking orders",
                "responses": {
                    "200": {
                        "description": "Booking orders encontrados",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.BookingOrder"
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Crea un nuevo registro de booking order",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "booking-order"
                ],
                "summary": "Crear booking order",
                "parameters": [
                    {
                        "description": "Datos del booking order",
                        "name": "bookingOrder",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.BookingOrder"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Booking order creado exitosamente",
                        "schema": {
                            "$ref": "#/definitions/models.BookingOrder"
                        }
                    },
                    "400": {
                        "description": "Datos inválidos",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/booking-order/user/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Obtiene todos los registros de booking orders por ID de usuario",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "booking-order"
                ],
                "summary": "Obtener todos los booking orders por ID de usuario",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del usuario",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Booking orders encontrados",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.BookingOrder"
                            }
                        }
                    },
                    "400": {
                        "description": "ID inválido",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Booking order no encontrado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/booking-order/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Obtiene un registro de booking order por su ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "booking-order"
                ],
                "summary": "Obtener booking order por ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del booking order",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Booking order encontrado",
                        "schema": {
                            "$ref": "#/definitions/models.BookingOrder"
                        }
                    },
                    "400": {
                        "description": "ID inválido",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Booking order no encontrado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Actualiza un registro de booking order",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "booking-order"
                ],
                "summary": "Actualizar booking order",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del booking order",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos del booking order",
                        "name": "bookingOrder",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.updateBookingOrderReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Booking order actualizado exitosamente",
                        "schema": {
                            "$ref": "#/definitions/models.BookingOrder"
                        }
                    },
                    "400": {
                        "description": "Datos inválidos",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.updateBookingOrderReq": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "paymentProviderId": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/models.PaymentStatus"
                }
            }
        },
...
        "models.BookingOrder": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "eventHour": {
                    "type": "string"
                },
                "eventName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Seat"
                    }
                },
                "paymentProviderId": {
                    "description": "Token o ID de transacción de la pasarela de pago (Stripe/MercadoPago)",
                    "type": "string"
                },
                "seatIds": {
                    "description": "Asientos involucrados en esta orden\nSeatIDs []string `gorm:\"type:text[]\" json:\"seatIds\"`",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "status": {
                    "$ref": "#/definitions/models.PaymentStatus"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
...
        "models.PaymentStatus": {
            "type": "string",
            "enum": [
                "PENDING",
                "COMPLETED",
                "FAILED"
            ],
            "x-enum-varnames": [
                "PaymentPending",
                "PaymentCompleted",
                "PaymentFailed"
            ]
        },
        "models.Seat": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "eventHour": {
                    "type": "string"
                },
                "eventId": {
                    "description": "Relaciones",
                    "type": "string"
                },
                "eventName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "lockedAt": {
                    "description": "Cuando se desbloquea",
                    "type": "string"
                },
                "lockedBy": {
                    "description": "Control de Bloqueo Temporal",
                    "type": "string"
                },
                "number": {
                    "description": "\"10\", \"A1\", etc.",
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "section": {
                    "description": "Identificación del asiento (ej: A-10, B-5)",
                    "type": "string"
                },
                "status": {
                    "description": "Estado actual del asiento\nenums: AVAILABLE, RESERVED, SOLD, BLOCKED",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.SeatStatus"
                        }
                    ]
                },
                "ticketId": {
                    "description": "ID del ticket final si se vende",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.SeatStatus": {
            "description": "Estado del asiento",
            "type": "string",
            "enum": [
                "AVAILABLE",
                "LOCKED",
                "SOLD"
            ],
            "x-enum-varnames": [
                "StatusAvailable",
                "StatusLocked",
                "StatusSold"
            ]
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Formato: Bearer \u003ctoken\u003e\nOperaciones para gestionar eventos\nOperaciones para gestionar asientos\nOperaciones para gestionar órdenes de reserva\nOperaciones para gestionar el proceso de checkout",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}